# App/Vacation/Resources/Public/swagger-ui/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "vacations",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/vacations/users/processVacationRequest/{id}.json": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Process Vacation Request for user",
                "description": "Process Vacation Request for user",
                "operationId": "processVacationRequest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "VacationRequest id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "label": "id"
                    },
                    {
                        "name": "entity[state]",
                        "in": "query",
                        "description": "State (0 = Pending, 1 = Approved, 2 = Denied)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": ["0", "1", "2"]
                        },
                        "label": "state"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "headers": {
                            "X-Rate-Limit": {
                                "description": "calls per hour allowed by the user",
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            },
                            "X-Expires-After": {
                                "description": "date in UTC when token expires",
                                "schema": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid athorization token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}